Split setting query things out into SettingQuery.scala
